Update: organized auth files and created a `layout.tsx` file for shared styles
