+++ src/app/(auth)/auth.module.scss
