+++ src/app/(auth)/register/page.tsx
@@ -1,38 +1,34 @@
 'use client'
 
-import { EmailIcon, LargeLogo, PasswordIcon } from '@/assets/icons'
+import { EmailIcon, PasswordIcon } from '@/assets/icons'
 
 import { Button } from '@/components'
-import { ILoginFormValues } from './login.types'
+import { IRegisterFormValues } from './register.types'
 import { Input } from '@/components'
 import Link from 'next/link'
-import React from 'react'
-import styles from './login.module.scss'
+import styles from '../auth.module.scss'
 import { useForm } from 'react-hook-form'
 
-const Login = () => {
+const Register = () => {
   const {
     register,
     formState: { errors },
     handleSubmit,
-    watch,
-  } = useForm<ILoginFormValues>()
+  } = useForm<IRegisterFormValues>()
 
-  const handleLoginSubmit = (data: ILoginFormValues) => {
+  const handleLoginSubmit = (data: IRegisterFormValues) => {
     console.log(data)
   }
 
   return (
-    <div className={styles.container}>
-      <LargeLogo />
       <form
         onSubmit={handleSubmit(handleLoginSubmit)}
         className={styles.container__auth_form}
       >
         <div className={styles.container__about_auth}>
-          <h3 className={styles.container__title}>Login</h3>
+          <h3 className={styles.container__title}>Create account</h3>
           <p className={styles.container__description}>
-            Add your details below to get back into the app
+            Let’s get you started sharing your links!
           </p>
         </div>
         <div className={styles.container__inputs}>
@@ -47,21 +43,30 @@ const Login = () => {
           <Input
             error={errors}
             leftIcon={<PasswordIcon />}
-            placeholder="Enter your password"
-            label="Password"
+            placeholder="At least 8 characters"
+            label="Create password"
             {...register('password', { required: 'Please check again' })}
           />
+          <Input
+            error={errors}
+            leftIcon={<PasswordIcon />}
+            placeholder="At least 8 characters"
+            label="Confirm password"
+            {...register('confirmPassword', { required: 'Please check again' })}
+          />
         </div>
+        <span className={styles.container__password_description}>
+          Password must contain at least 8 characters
+        </span>
         <Button variant="primary">Login</Button>
         <span className={styles.auth_description}>
-          Don't have an account?{' '}
-          <Link className={'link'} href="/register">
-            Create account
+          Already have an account?{' '}
+          <Link className={'link'} href="/login">
+            Login
           </Link>
         </span>
       </form>
-    </div>
   )
 }
 
-export default Login
+export default Register
