+++ src/app/(auth)/register/register.types.ts
